Conflicts:
	Makefile
	inav.creator.user
	inav.files
	lib/main/STM32F4xx_StdPeriph_Driver/Release_Notes.html
	lib/main/STM32F4xx_StdPeriph_Driver/inc/misc.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_adc.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_can.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_cryp.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_dac.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_dbgmcu.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_dcmi.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_dma.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_dma2d.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_exti.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_flash.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_fmc.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_fsmc.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_gpio.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_hash.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_i2c.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_iwdg.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_ltdc.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_pwr.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_rcc.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_rng.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_rtc.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_sai.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_sdio.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_spi.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_syscfg.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_tim.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_usart.h
	lib/main/STM32F4xx_StdPeriph_Driver/inc/stm32f4xx_wwdg.h
	lib/main/STM32F4xx_StdPeriph_Driver/src/misc.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_adc.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_can.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_crc.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_cryp.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_cryp_aes.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_cryp_des.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_cryp_tdes.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_dac.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_dbgmcu.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_dcmi.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_dma.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_dma2d.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_exti.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_flash.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_fmc.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_fsmc.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_gpio.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_hash.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_hash_md5.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_hash_sha1.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_i2c.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_iwdg.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_ltdc.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_pwr.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_rcc.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_rng.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_rtc.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_sai.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_sdio.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_spi.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_syscfg.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_tim.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_usart.c
	lib/main/STM32F4xx_StdPeriph_Driver/src/stm32f4xx_wwdg.c
	src/main/common/atomic.h
	src/main/config/config.c
	src/main/drivers/accgyro_mpu.c
	src/main/drivers/bus_spi.c
	src/main/drivers/bus_spi.h
	src/main/drivers/gpio.h
	src/main/drivers/inverter.c
	src/main/drivers/serial_uart.c
	src/main/drivers/serial_uart.h
	src/main/drivers/serial_usb_vcp.c
	src/main/drivers/sound_beeper.c
	src/main/drivers/timer.c
	src/main/drivers/timer.h
	src/main/main.c
	src/main/platform.h
	src/main/scheduler/scheduler_tasks.c
	src/main/startup/startup_stm32f40xx.s
	src/main/target/CC3D/target.h
	src/main/target/NAZE/target.h
	src/main/target/NUCLEOF7/target.h
	src/main/target/SPARKY/target.h
	src/main/target/stm32_flash_f405.ld

+++ src/main/drivers/serial_uart_hal.c
@@ -457,7 +457,7 @@ void uartStartTxDMA(uartPort_t *s)
 
 }
 
-uint8_t uartTotalRxBytesWaiting(serialPort_t *instance)
+uint32_t uartTotalRxBytesWaiting(serialPort_t *instance)
 {
     uartPort_t *s = (uartPort_t*)instance;
